requirements.txt
presidio-analyzer
presidio-anonymizer
spacy
pandas
tqdm
pyyaml>=6.0
openpyxl>=3.1